Move shared code and split into provider files.
