+++ src/native/minipal/CMakeLists.txt
@@ -13,6 +13,8 @@ set(SOURCES
     xoshiro128pp.c
     log.c
     env.c
+    env-system-environ-provider.c
+    env-android-system-properties-provider.c
 )
 
 # Provide an object library for scenarios where we ship static libraries
