+++ src/native/minipal/configure.cmake
@@ -5,23 +5,24 @@ include(CheckSymbolExists)
 check_include_files("windows.h;bcrypt.h" HAVE_BCRYPT_H)
 check_include_files("sys/auxv.h;asm/hwcap.h" HAVE_AUXV_HWCAP_H)
 check_include_files("asm/hwprobe.h" HAVE_HWPROBE_H)
-check_include_files(crt_externs.h HAVE_CRT_EXTERNS_H)
+check_include_files("crt_externs.h" HAVE_CRT_EXTERNS_H)
 
-check_function_exists(sysctlbyname HAVE_SYSCTLBYNAME)
-check_function_exists(fsync HAVE_FSYNC)
+check_symbol_exists(sysctlbyname "sys/sysctl.h" HAVE_SYSCTLBYNAME)
+check_symbol_exists(fsync "unistd.h" HAVE_FSYNC)
 
 check_symbol_exists(arc4random_buf "stdlib.h" HAVE_ARC4RANDOM_BUF)
-check_symbol_exists(O_CLOEXEC fcntl.h HAVE_O_CLOEXEC)
-check_symbol_exists(CLOCK_MONOTONIC time.h HAVE_CLOCK_MONOTONIC)
-check_symbol_exists(CLOCK_MONOTONIC_COARSE time.h HAVE_CLOCK_MONOTONIC_COARSE)
-check_symbol_exists(clock_gettime_nsec_np time.h HAVE_CLOCK_GETTIME_NSEC_NP)
+check_symbol_exists(O_CLOEXEC "fcntl.h" HAVE_O_CLOEXEC)
+check_symbol_exists(CLOCK_MONOTONIC "time.h" HAVE_CLOCK_MONOTONIC)
+check_symbol_exists(CLOCK_MONOTONIC_COARSE "time.h" HAVE_CLOCK_MONOTONIC_COARSE)
+check_symbol_exists(clock_gettime_nsec_np "time.h" HAVE_CLOCK_GETTIME_NSEC_NP)
 
-check_function_exists(sprintf_s HAVE_SPRINTF_S)
-check_function_exists(strncasecmp HAVE_STRNCASECMP)
-check_function_exists(strcpy_s HAVE_STRCPY_S)
-check_function_exists(strncpy_s HAVE_STRNCPY_S)
-check_function_exists(strcat_s HAVE_STRCAT_S)
-check_function_exists(getenv_s HAVE_GETENV_S)
+check_symbol_exists(vsprintf_s "stdio.h" HAVE_VSPRINTF_S)
+check_symbol_exists(strncasecmp "strings.h" HAVE_STRNCASECMP)
+check_symbol_exists(strcpy_s "string.h" HAVE_STRCPY_S)
+check_symbol_exists(strncpy_s "string.h" HAVE_STRNCPY_S)
+check_symbol_exists(strcat_s "string.h" HAVE_STRCAT_S)
+check_symbol_exists(getenv_s "stdlib.h" HAVE_GETENV_S)
+check_symbol_exists(getenv "stdlib.h" HAVE_GETENV)
 
 if (HAVE_CRT_EXTERNS_H)
     check_c_source_compiles(
@@ -31,6 +32,13 @@ if (HAVE_CRT_EXTERNS_H)
     "
     HAVE__NSGETENVIRON)
 endif()
+
+check_c_source_compiles(
+"
+#include <stdlib.h>
+int main(void) { char** e = _environ; return 0; }
+"
+HAVE__ENVIRON)
 
 if(CMAKE_C_BYTE_ORDER STREQUAL "BIG_ENDIAN")
     set(BIGENDIAN 1)
