+++ src/native/minipal/env.c
@@ -2,37 +2,15 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 
 #include "minipalconfig.h"
-#include "env.h"
-#include <stdlib.h>
 #include <errno.h>
 #include <assert.h>
-#include <stdio.h>
-#include <string.h>
 #include <ctype.h>
-#include "utils.h"
-#include "mutex.h"
-
-#if HAVE_CRT_EXTERNS_H
-#include <crt_externs.h>
-#endif
-
-typedef enum _EnvironmentVariableProviderType
-{
-    ENV_VAR_SYSTEM_ENVIRON_PROVIDER_TYPE,
-    ENV_VAR_ANDROID_SYSTEM_PROPERTIES_PROVIDER_TYPE
-} EnvironmentVariableProviderType;
-
-// Structure representing environment variable providers. Different platforms may support different
-// implementations to load environment variables from different sources.
-typedef struct _EnvironmentVariableProvider
-{
-    EnvironmentVariableProviderType type;
-    char* (*getenv_func)(const char* name);
-    void (*free_env_func)(char* value);
-    int (*getenv_s_func)(size_t *required_len, char *buffer, size_t buffer_len, const char *name);
-    char** (*get_environ_func)(void);
-    void (*free_environ_func)(char** values);
-} EnvironmentVariableProvider;
+#include <string.h>
+#include "minipal/env.h"
+#include "minipal/atomic.h"
+#include "minipal/utils.h"
+#include "minipal/strings.h"
+#include "minipal/mutex.h"
 
 /**
  * Structure representing a dynamic array of environment variable strings.
@@ -55,612 +33,27 @@ static bool g_env_loaded;
 // except when accessed from minipal_env_get_environ_unsafe.
 static EnvironmentVariables g_env;
 
-#if HAVE_SPRINTF_S
-#define env_sprintf_s sprintf_s
-#else
-#define env_sprintf_s snprintf
-#endif
-
-#if HAVE_STRNCASECMP
-#define env_strncasecmp strncasecmp
-#elif HOST_WINDOWS
-#define env_strncasecmp _strnicmp
-#else
-#error Missing strncasecmp implementation
-#endif
-
-#ifdef HOST_WINDOWS
-#define env_strdup _strdup
-#else
-#define env_strdup strdup
-#endif
-
-#if HAVE_STRCPY_S
-#define env_strcpy_s strcpy_s
-#else
-static int env_strcpy_s(char* dest, size_t destsz, const char* src)
-{
-    if (dest == NULL || src == NULL || destsz == 0)
-    {
-        if (dest && destsz > 0)
-        {
-            dest[0] = '\0';
-        }
-
-        return EINVAL;
-    }
-
-    size_t src_len = strlen(src);
-    if (src_len + 1 > destsz)
-    {
-        dest[0] = '\0';
-        return ERANGE;
-    }
-
-    memcpy(dest, src, src_len + 1);
-    return 0;
-}
-#endif
-
-#if HAVE_STRNCPY_S
-#define env_strncpy_s strncpy_s
-#else
-static int env_strncpy_s(char* dest, size_t destsz, const char* src, size_t count)
-{
-    if (dest == NULL || src == NULL || destsz == 0)
-    {
-        if (dest && destsz > 0)
-        {
-            dest[0] = '\0';
-        }
-
-        return EINVAL;
-    }
-
-    if (count >= destsz)
-    {
-        dest[0] = '\0';
-        return ERANGE;
-    }
-
-    size_t i = 0;
-    for (; i < count && i < destsz - 1 && src[i] != '\0'; ++i)
-    {
-        dest[i] = src[i];
-    }
-
-    dest[i] = '\0';
-    return 0;
-}
-#endif
-
-#if HAVE_STRCAT_S
-#define env_strcat_s strcat_s
-#else
-static int env_strcat_s(char* dest, size_t destsz, const char* src)
-{
-    if (dest == NULL || src == NULL || destsz == 0)
-    {
-        if (dest && destsz > 0)
-        {
-            dest[0] = '\0';
-        }
-
-        return EINVAL;
-    }
-
-    size_t dest_len = 0;
-    for (; dest_len < destsz; ++dest_len)
-    {
-        if (dest[dest_len] == '\0')
-        {
-            break;
-        }
-    }
-
-    if (dest_len == destsz)
-    {
-        dest[0] = '\0';
-        return ERANGE;
-    }
-
-    size_t src_len = strlen(src);
-    if (dest_len + src_len + 1 > destsz)
-    {
-        dest[0] = '\0';
-        return ERANGE;
-    }
-
-    memcpy(dest + dest_len, src, src_len + 1);
-    return 0;
-}
-#endif
-
-#if HAVE_GETENV_S
-#define env_getenv_s getenv_s
-#else
-static int env_getenv_s(size_t *required_len, char *buffer, size_t buffer_len, const char *name)
-{
-    if (name == NULL || required_len == NULL || (buffer == NULL && buffer_len > 0))
-    {
-        return EINVAL;
-    }
-
-    char *value = getenv(name);
-    size_t value_len = value ? strlen(value) : 0;
-
-    *required_len = value ? (value_len + 1) : 0;
-
-    if (buffer && buffer_len > 0)
-    {
-        if (!value)
-        {
-            buffer[0] = '\0';
-            return 0;
-        }
-
-        if (buffer_len < value_len + 1)
-        {
-            buffer[0] = '\0';
-            return ERANGE;
-        }
-
-        memcpy(buffer, value, value_len + 1);
-    }
-
-    return 0;
-}
-#endif
-
-#ifndef HOST_WINDOWS
-#include <stdatomic.h>
-#if ATOMIC_POINTER_LOCK_FREE != 2 || ATOMIC_INT_LOCK_FREE != 2
-#error C11 atomic pointer and int types must be lock free
-#endif
-#endif
-
-/**
- * @brief Atomically loads an int value with memory ordering guarantees.
- *
- * @param ptr   Pointer to the int to load.
- *
- * @return The loaded int value.
- */
-static int env_atomic_load_int(volatile int* value)
-{
-#ifdef HOST_WINDOWS
-    volatile int loaded_value = *value;
-	_ReadWriteBarrier();
-	return (int)loaded_value;
-#else
-    atomic_thread_fence(memory_order_seq_cst);
-    return (int)atomic_load((volatile atomic_int *)value);
-#endif
-}
-
-/**
- * @brief Atomically loads a pointer value with memory ordering guarantees.
- *
- * @param ptr   Pointer to the pointer to load.
- *
- * @return The loaded pointer value.
- */
-static void* env_atomic_load_ptr(volatile void** ptr)
-{
-#ifdef HOST_WINDOWS
-    volatile void* loaded_value = *ptr;
-	_ReadWriteBarrier();
-	return (void*)loaded_value;
-#else
-    atomic_thread_fence(memory_order_seq_cst);
-    return (void*)atomic_load((volatile _Atomic(void *) *)ptr);
-#endif
-}
-
-/**
- * @brief Atomically stores a pointer value with memory ordering guarantees.
- *
- * @param dest      Pointer to the destination pointer to store to.
- * @param value     The value to store.
- */
-static void env_atomic_store_ptr(volatile void** dest, void* value)
-{
-#ifdef HOST_WINDOWS
-    InterlockedExchangePointer((PVOID volatile *)dest, (PVOID)value);
-#else
-    atomic_store((volatile _Atomic(void *) *)dest, value);
-    atomic_thread_fence(memory_order_seq_cst);
-#endif
-}
-
-/**
- * @brief Atomically stores an int value with memory ordering guarantees.
- *
- * @param dest      Pointer to the destination int to store to.
- * @param value     The value to store.
- */
-static void env_atomic_store_int(volatile int* dest, int value)
-{
-#ifdef HOST_WINDOWS
-    _InterlockedExchange((LONG volatile *)dest, (LONG)value);
-#else
-    atomic_store((volatile atomic_int *)dest, value);
-    atomic_thread_fence(memory_order_seq_cst);
-#endif
-}
-
-/**
- * @brief Atomically compares and swaps a pointer value.
- *
- * @param dest  Pointer to the destination pointer to compare and swap.
- * @param exch  The value to store if the comparison succeeds.
- * @param comp  The value to compare against.
- *
- * @return The previous value of the pointer.
- */
-static void* env_atomic_cas_ptr(volatile void** dest, void* exch, void* comp)
-{
-#ifdef HOST_WINDOWS
-    return InterlockedCompareExchangePointer((PVOID volatile *)dest, (PVOID)exch, (PVOID)comp);
-#else
-    atomic_compare_exchange_strong((volatile _Atomic(void *) *)dest, &comp, exch);
-    atomic_thread_fence(memory_order_seq_cst);
-    return comp;
-#endif
-}
-
+// Environment providers array.
 #ifdef HOST_ANDROID
-#include <sys/system_properties.h>
-
-// Android system properties are checked once and cached in a global variable.
-// -1 means not initialized, 0 means disabled, 1 means enabled.
-static volatile int g_android_system_properties_enabled = -1;
-
-// Passed to __system_property_foreach callback.
-typedef struct _SystemPropertyValues
+extern EnvironmentProvider g_minipal_env_system_environ_provider;
+extern EnvironmentProvider g_minipal_env_android_system_properties_provider;
+static EnvironmentProvider* g_env_providers[] =
 {
-    int index;
-    int max;
-    char** values;
-} SystemPropertyValues;
-
-// Prefixes and names used by dotnet Android system properties.
-#define ENV_DOTNET_PREFIX "DOTNET_"
-#define ENV_DOTNET_PREFIX_LEN STRING_LENGTH(ENV_DOTNET_PREFIX)
-#define SYS_PROPS_DEBUG_PREFIX "debug."
-#define SYS_PROPS_DEBUG_PREFIX_LEN STRING_LENGTH(SYS_PROPS_DEBUG_PREFIX)
-#define SYS_PROPS_DOTNET_PREFIX "dotnet."
-#define SYS_PROPS_DOTNET_PREFIX_LEN STRING_LENGTH(SYS_PROPS_DOTNET_PREFIX)
-#define ENABLED_SYS_PROPS "enable_sys_props"
-#define ENABLED_SYS_PROPS_LEN STRING_LENGTH(ENABLED_SYS_PROPS)
-
-/**
- * @brief Retrieves the value of an Android system property with the specified name.
- *
- * @param name          The name of the Android system property to retrieve.
- * @param value         Buffer to receive the value of the Android system property.
- * @param value_len     Length of value in bytes, at least PROP_VALUE_MAX.
- *
- * @return Returns length if the property was found and read successfully, -1 otherwise.
- */
-static int get_android_system_property(const char* name, char* value, size_t value_len)
+    &g_minipal_env_system_environ_provider,
+    &g_minipal_env_android_system_properties_provider
+};
+#else
+extern EnvironmentProvider g_minipal_env_system_environ_provider;
+static EnvironmentProvider* g_env_providers[] =
 {
-    assert(value_len >= PROP_VALUE_MAX);
+    &g_minipal_env_system_environ_provider
+};
+#endif
 
-    int ret = -1;
-    const prop_info* info = __system_property_find(name);
-    if (info != NULL)
-    {
-        char tmp_name[PROP_NAME_MAX];
-        ret = __system_property_read(info, tmp_name, value);
-    }
-
-    return ret;
-}
-
-/**
- * @brief Determines whether Android system property should be enabled.
- *
- * @return true if Android system properties are enabled, false otherwise.
- */
-static bool is_android_system_properties_enabled(void)
-{
-    char value[PROP_VALUE_MAX];
-
-    int enabled = env_atomic_load_int(&g_android_system_properties_enabled);
-    if (enabled == -1)
-    {
-        int ret = get_android_system_property(SYS_PROPS_DEBUG_PREFIX SYS_PROPS_DOTNET_PREFIX ENABLED_SYS_PROPS, value, ARRAY_SIZE(value));
-        if (ret == -1)
-        {
-            ret = get_android_system_property(SYS_PROPS_DOTNET_PREFIX ENABLED_SYS_PROPS, value, ARRAY_SIZE(value));
-        }
-
-        if (ret == 1 && *value == '1')
-        {
-            enabled = 1;
-        }
-        else
-        {
-            enabled = 0;
-        }
-
-        env_atomic_store_int(&g_android_system_properties_enabled, enabled);
-    }
-
-    assert(enabled == 0 || enabled == 1);
-    return enabled == 1 ? true : false;
-}
-
-/**
- * @brief Callback function used to enumerate Android system properties.
- *
- * @param info      Pointer to the property info structure.
- * @param cookie    Pointer to user data (SystemPropertyValues).
- */
-static void android_system_property_callback(const prop_info* info, void* cookie)
-{
-    SystemPropertyValues* sys_props = (SystemPropertyValues*)cookie;
-    if (sys_props == NULL)
-    {
-        return;
-    }
-
-    char name[PROP_NAME_MAX];
-    char value[PROP_VALUE_MAX];
-
-    // API 26 use different read.
-    int ret = __system_property_read(info, name, value);
-    if (ret <= 0)
-    {
-        return;
-    }
-
-    const char* property_name_prefix = "";
-    const char* property_delimiter = "=";
-    const char* property_name = name;
-    const char* property_value = value;
-
-    if (env_strncasecmp(property_name, SYS_PROPS_DEBUG_PREFIX, SYS_PROPS_DEBUG_PREFIX_LEN) == 0)
-    {
-        property_name += SYS_PROPS_DEBUG_PREFIX_LEN;
-    }
-
-    if (env_strncasecmp(property_name, SYS_PROPS_DOTNET_PREFIX, SYS_PROPS_DOTNET_PREFIX_LEN) == 0)
-    {
-        property_name += SYS_PROPS_DOTNET_PREFIX_LEN;
-        if (env_strncasecmp(property_name, ENABLED_SYS_PROPS, ENABLED_SYS_PROPS_LEN) == 0)
-        {
-            return;
-        }
-
-        // Ending property name with a "." and a number is a way to handle large property name and value.
-        // Property ending with index 0 include property name, up to PROP_VALUE_MAX in length.
-        // Property ending with index 1 - n include values. All concatenated into final value.
-        // debug.dotnet.long_prop.0 = ThisIsReallyALongPropertyNameAbove32Chars
-        // debug.dotnet.long_prop.1 = SomePropertyNameValue
-        char* index_start = strrchr(name, '.');
-        if (index_start != NULL && isdigit(index_start[1]) == 0)
-        {
-            index_start = NULL;
-        }
-
-        if (index_start != NULL)
-        {
-            index_start++;
-
-            char* index_end = NULL;
-            long index = strtoul(index_start, &index_end, 10);
-            if (index_start == index_end || errno == ERANGE)
-            {
-                index_start = NULL;
-            }
-            else
-            {
-                if (index != 0)
-                {
-                    return;
-                }
-
-                *(index_start - 1) = '\0';
-            }
-        }
-
-        if (sys_props->values == NULL && *property_value != '\0')
-        {
-            sys_props->max++;
-            return;
-        }
-
-        if (sys_props->values != NULL && *property_value != '\0')
-        {
-            assert(sys_props->index < sys_props->max);
-
-            if (index_start != NULL)
-            {
-                property_name = property_value;
-                property_value = "";
-            }
-            else if (strchr(property_value, '=') != NULL)
-            {
-                property_name = property_value;
-                property_value = "";
-                property_delimiter = "";
-            }
-
-            if (env_strncasecmp(property_name, ENV_DOTNET_PREFIX, ENV_DOTNET_PREFIX_LEN) != 0)
-            {
-                property_name_prefix = ENV_DOTNET_PREFIX;
-            }
-
-            // property_name_prefix + property_name + property_delimiter + property_value + \0
-            size_t alloc_size = strlen(property_name_prefix) + strlen(property_name) + strlen(property_delimiter) + 1;
-            if (*property_value != '\0')
-            {
-                alloc_size += strlen(property_value);
-            }
-            else
-            {
-                alloc_size += PROP_VALUE_MAX;
-            }
-
-            sys_props->values[sys_props->index] = (char*)malloc(alloc_size * sizeof(char));
-            if (sys_props->values[sys_props->index] != NULL)
-            {
-                env_strcpy_s(sys_props->values[sys_props->index], alloc_size, property_name_prefix);
-                env_strcat_s(sys_props->values[sys_props->index], alloc_size, property_name);
-                env_strcat_s(sys_props->values[sys_props->index], alloc_size, property_delimiter);
-                env_strcat_s(sys_props->values[sys_props->index], alloc_size, property_value);
-
-                // If the property name is indexed, we need to append the values in index 1 - n.
-                if (index_start)
-                {
-                    int index = 1;
-                    char name2[PROP_NAME_MAX];
-                    char value2[PROP_VALUE_MAX];
-
-                    size_t used_size = strlen(sys_props->values[sys_props->index]) + 1;
-
-                    do
-                    {
-                        env_sprintf_s(name2, ARRAY_SIZE(name2), "%s.%d", name, index);
-                        ret = get_android_system_property(name2, value2, ARRAY_SIZE(value2));
-                        if (ret > 0)
-                        {
-                            if (used_size + ret <= alloc_size)
-                            {
-                                env_strcat_s(sys_props->values[sys_props->index], alloc_size, value2);
-                            }
-                            else
-                            {
-                                // Reallocate the buffer to fit the new value.
-                                alloc_size += ret;
-                                char* new_value = (char*)malloc(alloc_size * sizeof(char));
-                                if (new_value != NULL)
-                                {
-                                    env_strcpy_s(new_value, alloc_size, sys_props->values[sys_props->index]);
-                                    env_strcat_s(new_value, alloc_size, value2);
-                                    free(sys_props->values[sys_props->index]);
-                                    sys_props->values[sys_props->index] = new_value;
-                                }
-                            }
-
-                            used_size += ret;
-                            index++;
-                        }
-                    }
-                    while (ret > 0);
-                }
-
-                sys_props->index++;
-            }
-        }
-    }
-}
-
-/**
- * @brief Retrieves a Android system properties starting with "debug.dotnet" or "dotnet." prefix.
- *
- * @return Newly allocated string containing the value. Returns NULL if not found or if
-*  Android system properties are not enabled.
- *
- * @remarks
- * - Caller is responsible freeing the returned string.
- */
-static char* get_system_property(const char* name)
-{
-    assert(name != NULL);
-
-    if (!is_android_system_properties_enabled())
-    {
-        return NULL;
-    }
-
-    char* result = NULL;
-
-    const prop_info* info = __system_property_find(name);
-    if (info != NULL)
-    {
-        SystemPropertyValues value = { 0 };
-        char* data[1] = { 0 };
-
-        value.max = 1;
-        value.values = data;
-
-        android_system_property_callback(info, &value);
-        result = data[0];
-    }
-
-    return result;
-}
-
-/**
- * @brief Frees the memory allocated for a Android system properties.
- *
- * @param system_props Pointer to the Android system property strings to free.
- */
-static void free_system_property(char* system_prop)
-{
-    if (system_prop != NULL)
-    {
-        free(system_prop);
-    }
-}
-
-/**
- * @brief Retrieves all Android system properties using "debug.dotnet" or "dotnet." prefixes as an array of strings.
- *
- * @return Returns a pointer to an array of strings containing filtered Android system properties,
- * or NULL if none are found or Android system properties are not enabled.
- *
- * @remarks
- * - The returned array should be freed by calling free_system_properties.
- */
-static char** get_system_properties(void)
-{
-    if (!is_android_system_properties_enabled())
-    {
-        return NULL;
-    }
-
-    SystemPropertyValues sys_props = { 0 };
-    __system_property_foreach(android_system_property_callback, &sys_props);
-    if (sys_props.max == 0)
-    {
-        return NULL;
-    }
-
-    sys_props.values = (char**)malloc((sys_props.max + 1) * sizeof(char*));
-    if (sys_props.values != NULL)
-    {
-        memset(sys_props.values, 0, (sys_props.max + 1) * sizeof(char*));
-        sys_props.index = 0;
-        __system_property_foreach(android_system_property_callback, &sys_props);
-    }
-
-    return sys_props.values;
-}
-
-/**
- * @brief Frees the memory allocated for the array of Android system properties.
- *
- * @param system_props Pointer to the array of Android system property strings to free.
- */
-static void free_system_properties(char** system_props)
-{
-    if (system_props != NULL)
-    {
-        for (int i = 0; system_props[i] != NULL; i++)
-        {
-            free(system_props[i]);
-        }
-        free(system_props);
-    }
-}
-#endif // HOST_ANDROID
+// Number of environment providers.
+#define g_env_providers_count ARRAY_SIZE(g_env_providers)
 
 // Forward declares.
-static char** get_system_environ(void);
 static minipal_mutex* atomic_load_env_lock_ptr(void);
 static bool atomic_cas_env_lock_ptr(minipal_mutex* lock);
 static bool lock_env(void);
@@ -682,63 +75,14 @@ static bool foreach_env(EnvironmentVariables* env, bool (*callback)(const char* 
 static bool load_environ(void);
 static void unload_environ(void);
 
-// Environment variable providers array.
-static EnvironmentVariableProvider g_env_providers[] =
-{
-    {
-        ENV_VAR_SYSTEM_ENVIRON_PROVIDER_TYPE,
-        NULL,
-        NULL,
-        env_getenv_s,
-        get_system_environ,
-        NULL
-    }
-#ifdef HOST_ANDROID
-    ,
-    {
-        ENV_VAR_ANDROID_SYSTEM_PROPERTIES_PROVIDER_TYPE,
-        get_system_property,
-        free_system_property,
-        NULL,
-        get_system_properties,
-        free_system_properties
-    }
-#endif
-};
-
-// Number of environment variable providers.
-#define g_env_providers_count ARRAY_SIZE(g_env_providers)
-
-/**
- * @brief Retrieves the system environment variables array.
- *
- * @return Pointer to the system environment variables array.
- *
- * @remarks
- * - Returned system environment variables array is not thread-safe.
- */
-static char** get_system_environ(void)
-{
-    char** sys_env;
-
-#if HAVE__NSGETENVIRON
-    sys_env = *(_NSGetEnviron());
-#else   // HAVE__NSGETENVIRON
-    extern char **environ;
-    sys_env = environ;
-#endif  // HAVE__NSGETENVIRON
-
-    return sys_env;
-}
-
 /**
  * @brief Atomically loads the pointer to the global environment mutex lock.
  *
  * @return Pointer to the global environment mutex lock.
  */
-static minipal_mutex* atomic_load_env_lock_ptr()
+static minipal_mutex* atomic_load_env_lock_ptr(void)
 {
-    return (minipal_mutex *)env_atomic_load_ptr((volatile void **)&g_env_lock);
+    return (minipal_mutex *)minipal_atomic_load_ptr((volatile void **)&g_env_lock);
 }
 
 /**
@@ -754,7 +98,7 @@ static minipal_mutex* atomic_load_env_lock_ptr()
 static bool atomic_cas_env_lock_ptr(minipal_mutex* lock)
 {
     assert(lock != NULL);
-    return (env_atomic_cas_ptr((volatile void **)&g_env_lock, (void *)lock, NULL) == NULL);
+    return (minipal_atomic_cas_ptr((volatile void **)&g_env_lock, (void *)lock, NULL) == NULL);
 }
 
 /**
@@ -831,12 +175,12 @@ static bool resize_env(EnvironmentVariables* env, int new_size)
             if (env->data != NULL)
             {
                 char**old_env = env->data;
-                env_atomic_store_ptr((volatile void**)&env->data, NULL);
+                minipal_atomic_store_ptr((volatile void**)&env->data, NULL);
                 free(old_env);
             }
 
             env->capacity = new_size;
-            env_atomic_store_ptr((volatile void**)&env->data, (void*)new_env);
+            minipal_atomic_store_ptr((volatile void**)&env->data, (void*)new_env);
 
             result = true;
         }
@@ -857,7 +201,7 @@ static void destroy_env(EnvironmentVariables* env)
 
     char** data = env->data;
 
-    env_atomic_store_ptr((volatile void**)&env->data, NULL);
+    minipal_atomic_store_ptr((volatile void**)&env->data, NULL);
     env->capacity = 0;
     env->size = 0;
 
@@ -918,9 +262,9 @@ static EnvironmentVariables init_env(void)
 
     for (int i = 0; i < g_env_providers_count; ++i)
     {
-        if (g_env_providers[i].get_environ_func != NULL)
+        if (g_env_providers[i]->get_environ_func != NULL)
         {
-            env_provider_data[i] = g_env_providers[i].get_environ_func();
+            env_provider_data[i] = g_env_providers[i]->get_environ_func();
             if (env_provider_data[i] != NULL)
             {
                 int data_index = 0;
@@ -948,13 +292,13 @@ static EnvironmentVariables init_env(void)
                 {
                     if (i == 0)
                     {
-                        env.data[env.size] = env_strdup(env_provider_data[i][data_index]);
+                        env.data[env.size] = minipal_strdup(env_provider_data[i][data_index]);
                         data_index++;
                         env.size++;
                     }
                     else
                     {
-                        put_env(&env, env_strdup(env_provider_data[i][data_index]));
+                        put_env(&env, minipal_strdup(env_provider_data[i][data_index]));
                         data_index++;
                     }
 
@@ -967,9 +311,9 @@ static EnvironmentVariables init_env(void)
 
     for (int i = 0; i < g_env_providers_count; ++i)
     {
-        if (g_env_providers[i].free_environ_func != NULL && env_provider_data[i] != NULL)
+        if (g_env_providers[i]->free_environ_func != NULL && env_provider_data[i] != NULL)
         {
-            g_env_providers[i].free_environ_func(env_provider_data[i]);
+            g_env_providers[i]->free_environ_func(env_provider_data[i]);
         }
     }
 
@@ -1009,7 +353,7 @@ static bool copy_env(const EnvironmentVariables* src, EnvironmentVariables* dest
     {
         if (src->data[i] != NULL)
         {
-            dest->data[i] = env_strdup(src->data[i]);
+            dest->data[i] = minipal_strdup(src->data[i]);
             if (dest->data[i] == NULL)
             {
                 destroy_env(dest);
@@ -1043,7 +387,7 @@ static void assign_env(const EnvironmentVariables* src, EnvironmentVariables* de
 
     dest->capacity = src->capacity;
     dest->size = src->size;
-    env_atomic_store_ptr((volatile void**)&dest->data, (void*)src->data);
+    minipal_atomic_store_ptr((volatile void**)&dest->data, (void*)src->data);
 }
 
 /**
@@ -1085,7 +429,7 @@ static bool merge_envs(EnvironmentVariables* src, EnvironmentVariables* dest, bo
         }
         else
         {
-            put_env(dest, env_strdup(src->data[i]));
+            put_env(dest, minipal_strdup(src->data[i]));
         }
     }
 
@@ -1117,17 +461,17 @@ static const char* cache_env_string(EnvironmentVariables* env, const char* name)
 
     for (int i = g_env_providers_count; i > 0; --i)
     {
-        EnvironmentVariableProvider* provider = &g_env_providers[i - 1];
+        EnvironmentProvider* provider = g_env_providers[i - 1];
         if (provider->getenv_s_func != NULL)
         {
             size_t value_len;
             if (provider->getenv_s_func(&value_len, NULL, 0, name) == 0 && value_len != 0)
             {
-                size_t env_s_len = name_len + value_len + 1; // +1 for '=', env_getenv_s returns value_len including null terminator.
+                size_t env_s_len = name_len + value_len + 1; // +1 for '=', getenv_s_func returns value_len including null terminator.
                 env_s = (char*)malloc(env_s_len);
                 if (env_s != NULL)
                 {
-                    env_sprintf_s(env_s, env_s_len, "%s=", name);
+                    minipal_sprintf_s(env_s, env_s_len, "%s=", name);
                     provider->getenv_s_func(&value_len, env_s + name_len + 1, value_len, name);
                     break;
                 }
@@ -1142,10 +486,10 @@ static const char* cache_env_string(EnvironmentVariables* env, const char* name)
                 env_s = (char *) malloc(env_s_len);
                 if (env_s != NULL)
                 {
-                    env_sprintf_s(env_s, env_s_len, "%s=%s", name, value);
+                    minipal_sprintf_s(env_s, env_s_len, "%s=%s", name, value);
                 }
 
-                if (provider->free_env_func != NULL && value != NULL)
+                if (provider->free_env_func != NULL)
                 {
                     provider->free_env_func(value);
                 }
@@ -1284,7 +628,7 @@ static bool get_env_s(EnvironmentVariables* env, size_t* len, char* value, size_
         {
             if (value != NULL)
             {
-                env_strncpy_s(value, valuesz, env_value, env_value_len);
+                minipal_strncpy_s(value, valuesz, env_value, env_value_len);
                 value[env_value_len] = '\0';
             }
 
@@ -1297,7 +641,7 @@ static bool get_env_s(EnvironmentVariables* env, size_t* len, char* value, size_
         {
             if (value != NULL && valuesz > 0 && truncate)
             {
-                env_strncpy_s(value, valuesz, env_value, valuesz - 1);
+                minipal_strncpy_s(value, valuesz, env_value, valuesz - 1);
                 value[valuesz - 1] = '\0';
             }
 
@@ -1460,7 +804,7 @@ static bool set_env(EnvironmentVariables* env, const char* name, const char* val
         return false;
     }
 
-    env_sprintf_s(env_var, len, "%s=%s", name, value);
+    minipal_sprintf_s(env_var, len, "%s=%s", name, value);
     if (!put_env(env, env_var))
     {
         free(env_var);
@@ -1634,7 +978,7 @@ char** minipal_env_get_environ_unsafe(void)
 
     errno = 0;
 
-    result = (char**)env_atomic_load_ptr((volatile void **)&g_env.data);
+    result = (char**)minipal_atomic_load_ptr((volatile void **)&g_env.data);
 
     if (!result)
     {
@@ -1736,7 +1080,7 @@ char* minipal_env_get(const char* name)
     if (lock_env())
     {
         const char* value = get_env(&g_env, name);
-        result = value ? env_strdup(value) : NULL;
+        result = value ? minipal_strdup(value) : NULL;
         unlock_env();
     }
 
@@ -1833,7 +1177,7 @@ bool minipal_env_put(const char* env_s)
         return false;
     }
 
-    char* copy = env_strdup(env_s);
+    char* copy = minipal_strdup(env_s);
     if (copy == NULL)
     {
         errno = ENOMEM;
