+++ src/native/minipal/strings.h
@@ -35,6 +35,36 @@ CHAR16_T minipal_tolower_invariant(CHAR16_T code);
  */
 size_t minipal_u16_strlen(const CHAR16_T* str);
 
+/**
+ * @brief xplat implementation of sprintf_s.
+ */
+int minipal_sprintf_s(char* buffer, size_t count, const char* format, ...);
+
+/**
+ * @brief xplat implementation of strncasecmp.
+ */
+int minipal_strncasecmp(const char* str1, const char* str2, size_t n);
+
+/**
+ * @brief xplat implementation of strdup.
+ */
+char* minipal_strdup(const char *str);
+
+/**
+ * @brief xplat implementation of strcpy_s.
+ */
+int minipal_strcpy_s(char* dest, size_t destsz, const char* src);
+
+/**
+ * @brief xplat implementation of strncpy_s.
+ */
+int minipal_strncpy_s(char* dest, size_t destsz, const char* src, size_t count);
+
+/**
+ * @brief xplat implementation of strcat_s.
+ */
+int minipal_strcat_s(char* dest, size_t destsz, const char* src);
+
 #ifdef __cplusplus
 }
 #endif // __cplusplus
